backend/requirements.txt
networkx>=3.0
matplotlib>=3.6.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pymongo>=4.0.0
motor>=3.0.0
bcrypt>=4.0.0
pydantic[email]>=2.0.0
groq>=0.4.0
